tests/wpt/web-platform-tests/html/browsers/history/the-location-interface/location-prototype-setting.html

<!doctype html>
<meta charset=utf-8>
<title>[[SetPrototypeOf]] on a location object should return false</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  test(function() {
    var origProto = Object.getPrototypeOf(location);
    assert_throws(new TypeError, function() {
      Object.setPrototypeOf(location, {});
    });
    assert_throws(new TypeError, function() {
      location.__proto__ = {};
    });
    assert_false(Reflect.setPrototypeOf(location, {}));
    assert_equals(Object.getPrototypeOf(location), origProto);
  });
</script>
